+++ server/presenters/authenticationProvider.js
@@ -1,0 +1,14 @@
+// @flow
+import { AuthenticationProvider } from "../models";
+
+export default function present(
+  authenticationProvider: AuthenticationProvider
+) {
+  return {
+    id: authenticationProvider.id,
+    name: authenticationProvider.name,
+    createdAt: authenticationProvider.createdAt,
+    isEnabled: authenticationProvider.enabled,
+    isConnected: true,
+  };
+}
